Replace controller spec with feature spec
Controller specs are deprecated and frowned upon because they test too much implementation behavior. Convert to a feature spec where we still have access to sessions via Warden, which we need to test user_return_to logic in the LoginController. As part of this work remove referrer param logic from LoginController which was copypasta from another app and has not been discussed in H2.