Validate EDTF dates
Fixes #165

We do this using a new custom validator that uses the EDTF gem to determine if the `Work#created_edtf` value is indeed valid EDTF. Should this PR also include a change to the front-end (using the EDTF npm package) so that users know in advance if the field is valid? I assumed not but can add that.