Validate work types and subtypes
Fixes #114

As part of this work, we need two different sorts of validations: one that validates a model instance, and one that can be called from the controller context. We need the latter validation because the new work form is supplied these values via its params, so there is no model instance to validate at that time.