Fix file removal in work form bug
This was a head-scratcher. Adding a file to the work form and them adding keywords, related links, related works, or contributors *could* result in one or more files being removed. Why? Apparently because of our use of buttons as submit buttons instead of as button-type buttons. Note that the search field used to enter new keywords still does trigger a submit (and, thus, could prematurely deposit a work), but it does not remove any files. We should ticket that elsewhere.