Defer to contributor to have knowledge of its name and type
This small refactoring makes a few small tweaks to the Contributor model and DescriptionGenerator service that allow the latter to defer to the former regarding how its name is formatted (and what its type is). It also tightens up a type signature (though maybe that change is six-of-one-half-dozen-of-the-other as we all learn how to get value out of Sorbet) and moves away from a potentially misleading Ruby hash style where the keys appear as strings but are actually symbols. `{ "value": "foo" }["value"]` for instance evaluates to `nil` rather than `"foo"` which might be surprising, and the extraneous quotes are doing nothing for us. So, dump them.