Update work show page with state changes and new PURLs
Fixes #508

This commit makes the work show page more interactive by updating the display dynamically, via ActionCable, whenever the work undergoes state changes (moving from draft to depositing to deposited, etc.). It also updates the Purl link when the item gets a druid assigned to it. This makes for a more interactive experience, and prevents page reloads for updates. It also serves as a working example of how to use ActionCable to deliver user-facing features.

Question for @astridu: should we animate/highlight/flash these changes when they occur, or is it preferable to go with the more inobtrusive experience (see screen cap for current implementation).

Includes:
* Add an ActionCable channel for streaming work updates to clients
* Rename 'return' to 'reject' for consistent domain language across the app
* Add a Stimulus controller to the work show page to handle ActionCable interaction
* Create a new event service that makes sure every work state change is accompanied by an associated event and a broadcast sent to the WorkUpdatesChannel
* Allow the work model to return user-friendly labels for supported states