Entrance users with a spinner
Do this so they know that they should expect a scintilling update to their work any moment now...

I do not love this approach of stuffing HTML in a string in the model, but I also do not have a better approach in my quiver at the moment. Happy to take suggestions and amend.