Align collection state flow with work state flow
This prevents a collection from going from deposited immediately to depositing without going through version_draft first. The collection is only in this state momentarily, so it adds a single DB row write but otherwise does not effect the functionality of the application, which is why no test changes are happening here.