Use remote forms consistently across app
Consistency is good. In doing this work, I replaced a single usage of `form_for` with `form_with` so the app does not use a mix of local and remote forms, and removed the model argument to this call which is/was unused. This commit also explicitly marks all of our forms as remote forms for future-proofing, as we know an upcoming Rails 6.1 change will make `form_with` default back to local forms.

I found that the mediated deposit versioning spec was flappy locally because there can be two links on the page with the work title, so I changed the spec to look for the link only in the scope of the collection name (the deposits in progress section). To do this, I moved the collection name from a section header to a table caption. I do not notice any UI difference with this change.