Show returned and version draft works as in progress in dashboard
Fixes #847

This PR adds to the "in progress" section of the dashboard, which previously displayed only works in the first draft state, by also displaying works in the version draft and rejected states. As part of this work, rename the `drafts` attr of the DashboardPresenter to be more accurate, and extract a view component for this section of the dashboard.