Update the work types and subtypes per latest in design document
Connects to #716

This also includes work from #905 (if you would not mind making sure I got everything, @peetucket, your eyes are most welcome on this commit) which should now be subsumed by this. It also includes a new "General" section in the types to genres and resource types mappings to reflect the new "more types" approach to subtypes. And finally it includes new tests to ensure ongoing cohesion between the types YAML configs and the work type and subtypes in the application/