Allow works to have a small number of legacy licenses
Fixes #1047

This change is intended to help migrate Hydrus 1 data into Hydrus 2, and to display migrated works with their current (CC 3.0) licenses. It allows a small set of displayable licenses to show both in the work show page and on the work form. Before a new version of such a work may be deposited again, a new H2 license must be selected.

Includes

* Change license validation in work model
  * Expand the list of allowable license values at the model level such that it includes both selectable and displayable licensed. It also permits the license to be nil to allow draft works to be valid without a license. This also allows us to migrate CC3 licenses from Hydrus 1.0.