Don't raise errors when running feature specs
Instead respond with production like responses (e.g. 404 rather than ActiveRecord::RecordNotFound).

This fixes the build because the delete test can delete a record before the delete_button endpoint is called, which leads it to raise a RecordNotFound.  This is handled as expected if it returns a 404 instead.

This code was also used in the hydra_etd project