Allow collections to be decommissioned
Fixes #2629

This commit adds the ability to decommission collections if and only if all of their works have previously been decommissioned (or if they have no works). The approach taken in this commit mirrors that used for decommissioning works, with deviations where necessary due to design differences between collections and works.