Generate filesets sorted by their filenames
Connects to sul-dlss/sul-embed#1380

This commit supports work to render folder/directory path information in filenames within sul-embed. Currently, sul-embed respects the order of contentMetadata in public XML (which is generated from from Cocina). Meanwhile, files for a given work in H2 are ordered randomly, based partly on 1) the order in which they were uploaded, and 2) whether the files are staged or preserved. This commit establishes a consistent order for files when generating Cocina for an H2 work, and that order is ascending alpha based on filenames, meaning sul-embed can continue to respect the given order.