Address unpermitted params across various controller interactions
Fixes #3191

This commit adds `with_orcid` as a permitted parameter for works in the `WorksController`. As part of this work, configure the app to raise on unpermitted params in the test and dev envs, so we can better flag these before they get to production going forward, and then deal with all the failing tests that causes.