Limit maximum files to 250
Limit the maximum number of files that can be uploaded individually to 250.
Doing any more causes DropZone to behave erratically in Firefox and Chrome.

It does not appear to be feasible to change the error message that is
displayed when a file exceeds 10GB to use GB instead the less common MiB
units. But a message is displayed so that's good.

Fixes #3183