[SPIKE] Use axe-core-rspec and nokogiri gems to test accessibility
This commit uses the axe-core-spec gem along with the newly added selenium web driver to audit the accessibility of selected pages. It also uses nokogiri to validate HTML of selected pages. This is all done to help improve accessibility and discourage regressions over time.

Note that this is a spike to show some patterns. We would want to add more page coverage to these new feature specs once we're settled on a pattern we like.